Have JNIUtil exporrtResource() preserve filename
Most operating systems will extract the symbolic name of a shared
library from it's internal structure but on Windows the name
of the file that was passed to System.load() matters. This change
preserves the name of the name of the exported file in JNIUtil which
in turn enables users to distribute the full set of native libraries
in jars.